+++ EngineSource/Objects/Components/ParticleComponent.h
@@ -9,8 +9,6 @@ namespace Particles
 class ParticleComponent : public Component
 {
 	Particles::ParticleEmitter* Emitter = nullptr;
-	Vector3 Position;
-	Vector3 Rotation;
 public:
 	void Begin() override;
 	void Tick() override;
@@ -21,10 +19,4 @@ public:
 	bool GetActive();
 	bool GetIsFinished();
 	void Reset();
-
-
-	void SetRelativeRotation(Vector3 NewRotation);
-	Vector3 GetRelativeRotation();
-	void SetRelativePosition(Vector3 NewPos);
-	Vector3 GetRelativePosition();
 };