+++ EngineSource/Engine/EngineProperties.h
@@ -16,5 +16,5 @@ namespace Project
 	void OnLaunch();
 	extern const char* ProjectName;
 }
-#define VERSION_STRING "1.7.0"
+#define VERSION_STRING "1.7.1"
 #define OPENGL_MIN_REQUIRED_VERSION "GL_VERSION_4_2"