+++ EngineSource/Engine/OS.cpp
@@ -10,6 +10,11 @@
 #endif
 
 #if __linux__
+#include <unistd.h>
+#include <ios>
+#include <iostream>
+#include <fstream>
+#include <string>
 #include <Engine/Utility/StringUtility.h>
 #endif
 
@@ -61,8 +66,37 @@ void OS::SetConsoleWindowVisible(bool Visible)
 #if __linux__
 size_t OS::GetMemUsage()
 {
-	return 0;
+   using std::ios_base;
+   using std::ifstream;
+   using std::string;
+
+   // 'file' stat seems to give the most reliable results
+   //
+   ifstream stat_stream("/proc/self/stat",ios_base::in);
+
+   // dummy vars for leading entries in stat that we don't care about
+   //
+   string pid, comm, state, ppid, pgrp, session, tty_nr;
+   string tpgid, flags, minflt, cminflt, majflt, cmajflt;
+   string utime, stime, cutime, cstime, priority, nice;
+   string O, itrealvalue, starttime;
+
+   // the two fields we want
+   //
+   size_t vsize;
+   long rss;
+
+   stat_stream >> pid >> comm >> state >> ppid >> pgrp >> session >> tty_nr
+               >> tpgid >> flags >> minflt >> cminflt >> majflt >> cmajflt
+               >> utime >> stime >> cutime >> cstime >> priority >> nice
+               >> O >> itrealvalue >> starttime >> vsize >> rss; // don't care about the rest
+
+   stat_stream.close();
+
+   long page_size_kb = sysconf(_SC_PAGE_SIZE) / 1024; // in case x86-64 is configured to use 2MB pages
+   return vsize / 1024.0;
 }
+
 
 std::wstring OS::Utf8ToWstring(std::string utf8)
 {
@@ -129,6 +163,8 @@ std::string OS::ShowOpenFileDialog()
 #endif
 
 #if __linux__
+
+// I <3 Linux
 std::string OS::ShowOpenFileDialog()
 {
 	char filename[4096];
@@ -139,7 +175,6 @@ std::string OS::ShowOpenFileDialog()
 		return "";
 	}
 	pclose(f);
-	Log::Print(filename);
 	std::string FilenameString = filename;
 	StrUtil::ReplaceChar(FilenameString, '\n', "");
 	return FilenameString;