+++ CHANGES.md
@@ -3,8 +3,12 @@
 ## Version 1.7.0
 
 ### Changes
-- Added Controller Input support. (Engine/Gamepad.h)
+- Added controller input support. (Engine/Gamepad.h)
+	- Added C# bindings for controller input.
+- Added Linux support.
+- Added pre-built binaries for Linux (maybe?)
 
 ### Fixes
+- Fixed some issues with point lights.
 - Fixed isues with shadows.
 - Fixed an issue with the default movement.