+++ CHANGES.md
@@ -12,3 +12,4 @@
 - Fixed some issues with point lights.
 - Fixed isues with shadows.
 - Fixed an issue with the default movement.
+-